refactor: 엣지 케이스 레이어 분리
