+++ src/utils/repeat/constants.ts
@@ -1,0 +1,3 @@
+export const MS_PER_DAY = 1000 * 60 * 60 * 24;
+export const MS_PER_WEEK = MS_PER_DAY * 7;
+export const GLOBAL_REPEAT_CAP = new Date('2025-10-30T00:00:00Z');
